+++ tests/test_util.py
@@ -33,3 +33,22 @@ def test_sanitize_url(url, expected):
 
 def test_url_without_fragment():
     assert url_without_fragment('http://example.org/p#frag#ment') == 'http://example.org/p'
+
+
+def test_metadata2markdown(tmp_path):
+    from pycldf import Generic
+
+    ds = Generic.in_dir(tmp_path)
+    ds.add_component('LanguageTable')
+    ds.add_sources("@book{b,\ntitle={the book}}")
+    ds.write(
+        LanguageTable=[dict(ID='l', Name='the language')],
+        zipped={'LanguageTable', 'sources.bib'}
+    )
+    md = metadata2markdown(ds, tmp_path / 'Generic-metadata.json')
+    assert 'sources.bib.zip' in md
+    assert 'languages.csv.zip' in md
+
+    tmp_path.joinpath('languages.csv.zip').unlink()
+    md = metadata2markdown(ds, tmp_path / 'Generic-metadata.json')
+    assert 'languages.csv.zip' not in md, "Don't link non-existing files"